+++ tests/Clear-AzDataTable.Tests.ps1
@@ -1,13 +1,4 @@
-param(
-    [Parameter()]
-    [ValidateScript({ $_ -match '\.psd1$' }, ErrorMessage = 'Please input a .psd1 file')]
-    $Manifest
-)
-
 BeforeDiscovery {
-    . "$PSScriptRoot\CommonTestLogic.ps1"
-    Invoke-ModuleReload -Manifest $Manifest
-
     # Get command from current test file name
     $Command = Get-Command ((Split-Path $PSCommandPath -Leaf) -replace '.Tests.ps1')
     $ParameterTestCases = @(
@@ -22,7 +13,7 @@ BeforeDiscovery {
     )
 }
 
-Describe 'New-AzDataTable' {
+Describe 'Clear-AzDataTable' {
     Context 'parameters' {
 
         It 'only has expected parameters' -TestCases @{ Command = $Command ; Parameters = $ParameterTestCases.Name } {