+++ tests/Remove-AzDataTableEntity.Tests.ps1
@@ -1,24 +1,7 @@
-param(
-    [Parameter()]
-    [ValidateScript({ $_ -match '\.psd1$' }, ErrorMessage = 'Please input a .psd1 file')]
-    $Manifest
-)
-
 BeforeDiscovery {
-    . "$PSScriptRoot\CommonTestLogic.ps1"
-    Invoke-ModuleReload -Manifest $Manifest
-
     # Get command from current test file name
     $Command = Get-Command ((Split-Path $PSCommandPath -Leaf) -replace '.Tests.ps1')
     $ParameterTestCases += @(
-        @{
-            Command       = $Command
-            Name          = 'Context'
-            Type          = 'PipeHow.AzBobbyTables.AzDataTableContext'
-            ParameterSets = @(
-                @{ Name = '__AllParameterSets'; Mandatory = $true }
-            )
-        }
         @{
             Command       = $Command
             Name          = 'Entity'
@@ -27,10 +10,18 @@ BeforeDiscovery {
                 @{ Name = '__AllParameterSets'; Mandatory = $true }
             )
         }
+        @{
+            Command       = $Command
+            Name          = 'Context'
+            Type          = 'PipeHow.AzBobbyTables.AzDataTableContext'
+            ParameterSets = @(
+                @{ Name = '__AllParameterSets'; Mandatory = $true }
+            )
+        }
     )
 }
 
-Describe 'Update-AzDataTableEntity' {
+Describe 'Remove-AzDataTableEntity' {
     Context 'parameters' {
 
         It 'only has expected parameters' -TestCases @{ Command = $Command ; Parameters = $ParameterTestCases.Name } {